feat: allow writing to previously disabled interval
There were two main issues preventing this feature that were addressed. The first is that the zookeeper nodes that reference disabled segments take time to drop; instead of waiting until this happens, they are now removed when delete_datasource is called. This got us a little further, but the second issue arose after implementing the first part. The task that was generated sticks around in Druid, when the next task is generated for the same interval, there was a conflict because it thought the task still existed. To circumvent this, I needed to enable random task ids.